Workaround uart_poll_out when muxing
Since uart_poll_out only passes a single character to the uart driver
and uart_mux is implemented as a decorating uart driver we end up
wrapping single characters in the GSM MUX frames.

This is a problem for the SIM7600E which ends up rebooting due to
bombardment of GSM MUX frames.