net: lwm2m: Make lwm2m_pull_context reusable
Add a semaphore to control that no collisions occur when multiple
sources want to use the pull_context

Add struct firmware_pull_context *ctx as an argument to the result_cb of
the context. This allows the receiver to do some kind of differentiation
on the source.

In order to not have to create one socket callback per user of the pull
context the signature of fault_cb was changed to include the lwm2m_ctx
which holds the callback pointer.

The find_msg helper function in lwm2m_engine.c is made non static in
order to get the context in the do_firmware_transfer_reply_cb.

Signed-off-by: Viktor Sjölind <viktor.sjolind@endian.se>