net: lwm2m: Pull context remove use of sys q
Match upstream changes in lwm2m_obj_firmware_pull.c removing the use of
the system work queue.

A side effect of this change is that multiple consumers of pull context
can download at the same time. This is bounded by number of UDP sockets
allowed at the same time which I guess is good since we do not crash but
it is really bad in terms of UX and API as it means that the order in
which one starts downloads affects which will be successful.

Also removes all use of global state in lwm2m_pull_context.c.

In order to not have to create one socket callback per user of the pull
context the signature of fault_cb was changed to include the lwm2m_ctx
which holds the callback pointer.

The find_msg helper function in lwm2m_engine.c is made non static in
order to get the context in the do_firmware_transfer_reply_cb.